Sexto Commit

+++ ExercicioLoja/Entidades/Pedido.cs
@@ -11,7 +11,7 @@ namespace ExercicioLoja.Entidades
         public virtual int Id { get; set; }
         public virtual DateTime DataDoPedido { get; set; }
         public virtual IList<Produto> Produtos { get; set; }
-        public virtual int ValorTotal { get; set; }
+        public virtual decimal ValorTotal { get; set; }
         public virtual Cliente Cliente { get; set; }            
     }
 }
